+++ lib/router.js
@@ -330,7 +330,7 @@ router.get('/qutoutiao/category/:cid', lazyloadRouteHandler('./routes/qutoutiao/
 // router.get('/bbc/:site?/:channel?', lazyloadRouteHandler('./routes/bbc/index'));
 
 // 看雪
-router.get('/pediy/topic/:category?/:type?', lazyloadRouteHandler('./routes/pediy/topic'));
+// router.get('/pediy/topic/:category?/:type?', lazyloadRouteHandler('./routes/pediy/topic'));
 
 // 老司机
 router.get('/laosiji/hot', lazyloadRouteHandler('./routes/laosiji/hot'));
