+++ lib/routes/kanxue/namespace.ts
@@ -1,0 +1,6 @@
+import type { Namespace } from '@/types';
+
+export const namespace: Namespace = {
+    name: '看雪',
+    url: 'kanxue.com',
+};
